Changes after editorial review.
