+++ Application/EdFi.Ods.Common/Security/Authorization/IAuthorizationStrategy.cs
@@ -16,17 +16,6 @@ namespace EdFi.Ods.Common.Security.Authorization
     /// </summary>
     public interface IAuthorizationStrategy
     {
-        // /// <summary>
-        // /// Authorize the request for a single item.
-        // /// </summary>
-        // /// <param name="relevantClaims">The subset of the caller's claims that are relevant for the authorization decision.</param>
-        // /// <param name="authorizationContext">The authorization context, including the entity's data.</param>
-        // /// <param name="cancellationToken"></param>
-        // Task AuthorizeSingleItemAsync(
-        //     IEnumerable<Claim> relevantClaims, 
-        //     EdFiAuthorizationContext authorizationContext, 
-        //     CancellationToken cancellationToken);
-
         /// <summary>
         /// Applies filtering to a multiple-item request.
         /// </summary>
