+++ Application/EdFi.Ods.Tests/EdFi.Ods.Security/AuthorizationStrategies/Relationships/AllRelationshipsAuthorizationStrategyTests.cs
@@ -377,97 +377,5 @@ namespace EdFi.Ods.Tests.EdFi.Ods.Api.Security.AuthorizationStrategies.Relations
                     .MustNotHaveHappened();
             }
         }
-
-        // TODO: GKM - Identify a redundant Postman test for this scenario.
-        // public class When_authorizing_a_single_item_request_with_claim_needed_for_local_request_authorization_but_EdOrg_values_dont_match
-        //     : ScenarioFor<RelationshipsWithEdOrgsAndPeopleAuthorizationStrategy<RelationshipsAuthorizationContextData>>
-        // {
-        //     private InstanceAuthorizationResult _actualResult;
-        //
-        //     private class TestEntity
-        //     {
-        //         public TestEntity(int localEducationAgencyId)
-        //         {
-        //             LocalEducationAgencyId = localEducationAgencyId;
-        //         }
-        //
-        //         public int LocalEducationAgencyId { get; }
-        //     }
-        //
-        //     protected override void Arrange()
-        //     {
-        //         Supplied(
-        //             "entity",
-        //             new TestEntity(
-        //                 Supplied("LocalEducationAgencyId", 999)));
-        //
-        //         Supplied(
-        //             new RelationshipsAuthorizationContextData
-        //             {
-        //                 LocalEducationAgencyId = Supplied<int>("LocalEducationAgencyId")
-        //             });
-        //
-        //         Given<education_organization_cache>()
-        //             .that_always_returns_a_Local_Education_Agency_for(Supplied<int>("LocalEducationAgencyId"))
-        //             .that_always_returns_a_Local_Education_Agency_for(777);
-        //
-        //         Given<IEducationOrganizationIdHierarchyProvider>()
-        //             .that_always_returns_an_empty_graph();
-        //
-        //         Given<context_data_transformer>(
-        //             new passthrough_context_data_transformer());
-        //
-        //         Given<context_data_provider>()
-        //             .that_returns_property_names(
-        //                 Supplied(
-        //                     "propertyNames",
-        //                     new[]
-        //                     {
-        //                         "LocalEducationAgencyId"
-        //                     }))
-        //             .that_given_entity(Supplied("entity"))
-        //             .Returns(Supplied<RelationshipsAuthorizationContextData>());
-        //
-        //         Given<context_data_provider_factory>()
-        //             .that_always_returns(Given<context_data_provider>());
-        //
-        //         Supplied(
-        //             Given_a_claim_for_an_arbitrary_resource_for_EducationOrganization_identifiers(777));
-        //
-        //         var fakeNamesProvider = A.Fake<IEducationOrganizationIdNamesProvider>();
-        //         A.CallTo(() => fakeNamesProvider.GetAllNames()).Returns(new[] { "LocalEducationAgencyId" });
-        //         Supplied(fakeNamesProvider);
-        //
-        //         var apiKeyContextProvider = A.Fake<IApiKeyContextProvider>();
-        //         Supplied(apiKeyContextProvider);
-        //     }
-        //
-        //     protected override void Act()
-        //     {
-        //         var filtering = SystemUnderTest.GetAuthorizationStrategyFiltering(
-        //             new[] { Supplied<Claim>() },
-        //             Given_an_authorization_context_with_entity_data(Supplied("entity")));
-        //
-        //         var filterDefinitions = (new RelationshipsAuthorizationStrategyFilterDefinitionsProvider(
-        //             Supplied<IEducationOrganizationIdNamesProvider>(),
-        //             Supplied<IApiKeyContextProvider>())).GetAuthorizationFilterDefinitions();
-        //
-        //         const string FILTER_NAME = "ClaimEducationOrganizationIdsToLocalEducationAgencyId";
-        //         
-        //         var filterDefinition = filterDefinitions.SingleOrDefault(
-        //             fd => fd.FilterName == FILTER_NAME); //filtering.Filters.Single().FilterName);
-        //
-        //         _actualResult = filterDefinition.AuthorizeInstance(
-        //             Given_an_authorization_context_with_entity_data(Supplied("entity")),
-        //             filtering.Filters.Single(f => f.FilterName == FILTER_NAME));
-        //     }
-        //
-        //     [Assert]
-        //     public void Should_throw_an_EdFiSecurityException_indicating_the_claims_did_not_provide_authorization_for_the_request()
-        //     {
-        //         _actualResult.Exception.ShouldBeExceptionType<EdFiSecurityException>();
-        //         _actualResult.Exception.MessageShouldContain("Access to the requested");
-        //     }
-        // }
     }
 }
